requirements.txt
Django==5.0.1
django-debug-toolbar==4.2.0
django-environ==0.11.2
djangorestframework==3.14.0
psycopg2-binary==2.9.9
